refactor: 스토리북 글로벌 스타일 설정 #19
